Trying to create a more modularized (object-oriented) version of the s plotter app because, generally, I think it makes more sense. For now, this is in the test_file section. I plan later to basically over-write the s-plotter application with this test_file upload application. There are several things that I like about this new approach: 	1. The html actually creates the initial version of the graph without any data. This makes it easier to understand. 	2. The plot, itself, behaves like an object which can be maniuplated. Again, easier to understand and easier to scale.
However, in the plot_data function, I cannot add the legend in the same way that it is added in the original s-plotter application.
I do not know why. The same is true of the paths for the actual data. There is some basic misunderstanding that I have about
how paths are appended to the actual svg--in this case the svg is graph.vis. So there is some homework to be done here.